+++ packages/terra-doc-template/src/terra-dev-site/doc/doc-template/Placeholder.2.doc.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import DocTemplate from '../../../DocTemplate';
 import ReadMe from '../../../../docs/PLACEHOLDER.md';
-import { name } from '../../../../package.json';
+import { name, version } from '../../../../package.json';
 
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 // Component Source
@@ -17,6 +17,7 @@ import PlaceholderLightSrc from '!raw-loader!../../../../src/terra-dev-site/doc/
 const DocPage = () => (
   <DocTemplate
     packageName={name}
+    packageVersion={version}
     readme={ReadMe}
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
     examples={[