+++ packages/terra-form-select/CHANGELOG.md
@@ -6,6 +6,10 @@ Unreleased
 ### Added
 * Added AT announcement for onClick and backspace in MultiSelect and TagSelect
 
+### Changed
+* Import for package version from package.json added to doc page
+
+
 5.26.0 - (August 14, 2019)
 ------------------
 ### Removed