+++ packages/terra-grid/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Import for package version from package.json added to doc page
 
 6.7.0 - (August 14, 2019)
 ------------------